# Conflicts:
#	facilities/src/main/java/gov/va/api/lighthouse/facilities/CmsOverlayControllerV0.java
#	facilities/src/main/java/gov/va/api/lighthouse/facilities/CmsOverlayControllerV1.java
#	facilities/src/main/java/gov/va/api/lighthouse/facilities/CmsOverlayEntity.java
#	facilities/src/main/java/gov/va/api/lighthouse/facilities/CmsOverlayHelper.java
#	facilities/src/main/java/gov/va/api/lighthouse/facilities/FacilityOverlayV0.java
#	facilities/src/main/java/gov/va/api/lighthouse/facilities/InternalFacilitiesController.java
#	facilities/src/main/java/gov/va/api/lighthouse/facilities/collector/CmsOverlayCollector.java
#	facilities/src/main/java/gov/va/api/lighthouse/facilities/collector/FacilitiesCollector.java
#	facilities/src/test/java/gov/va/api/lighthouse/facilities/CmsOverlayControllerV0Test.java
#	facilities/src/test/java/gov/va/api/lighthouse/facilities/CmsOverlayControllerV1Test.java
#	facilities/src/test/java/gov/va/api/lighthouse/facilities/CmsOverlayTransformerV0Test.java
#	facilities/src/test/java/gov/va/api/lighthouse/facilities/InternalFacilitiesControllerTest.java

+++ facilities/src/main/java/gov/va/api/lighthouse/facilities/InternalFacilitiesController.java
@@ -220,6 +220,7 @@ public class InternalFacilitiesController {
       log.info("Deleting cms overlay for id: {}", sanitize(id));
       overlayEntity.cmsOperatingStatus(null);
       overlayEntity.cmsServices(null);
+      overlayEntity.healthCareSystem(null);
     } else if (thisNodeOnly.equalsIgnoreCase("operating_status")) {
       if (overlayEntity.cmsOperatingStatus() == null) {
         log.info("CmsOverlay {} does not have an operating_status, ignoring request", sanitize(id));
@@ -234,11 +235,20 @@ public class InternalFacilitiesController {
       }
       log.info("Deleting detailed_services node from overlay for id: {}", sanitize(id));
       overlayEntity.cmsServices(null);
+    } else if (thisNodeOnly.equalsIgnoreCase("system")) {
+      if (overlayEntity.healthCareSystem() == null) {
+        log.info("CmsOverlay {} does not have health care system, ignoring request", sanitize(id));
+        return ResponseEntity.accepted().build();
+      }
+      log.info("Deleting health care system node from overlay for id: {}", sanitize(id));
+      overlayEntity.healthCareSystem(null);
     } else {
       log.info("CmsOverlay field {} does not exist.", sanitize(thisNodeOnly));
       throw new ExceptionsUtils.NotFound(thisNodeOnly);
     }
-    if (overlayEntity.cmsOperatingStatus() == null && overlayEntity.cmsServices() == null) {
+    if (overlayEntity.cmsOperatingStatus() == null
+        && overlayEntity.cmsServices() == null
+        && overlayEntity.healthCareSystem() == null) {
       cmsOverlayRepository.delete(overlayEntity);
     } else {
       cmsOverlayRepository.save(overlayEntity);
